pyproject.toml
[project]
name = "py-test-uv-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=7.1.0",
    "jupyter-client>=8.6.3",
    "plotly-express>=0.4.1",
    "polars[pyarrow]>=1.34.0",
    "psutil>=7.1.3",
    "pyarrow>=22.0.0",
    "pyomo>=6.9.5",
    "tornado>=6.5.2",
    "traitlets>=5.14.3",
]
